+++ vecorel_cli/vecorel/extensions.py
@@ -1,0 +1,3 @@
+ADMIN_DIVISION = "https://vecorel.org/administrative-division-extension/v0.1.0/schema.yaml"
+
+GEOMETRY_METRICS = "https://vecorel.org/geometry-metrics-extension/v0.1.0/schema.yaml"